MergeSort

+++ algorithms/sort/insertion_sort.h
@@ -1,12 +1,14 @@
 #ifndef SORT_H
 #define SORT_H
 
+#include <iostream>
+
 template<class T>
-void insertion_sort(T a[], int length)
+void insertion_sort(T a[], size_t length)
 {
     int key;
     int j;
-    for(int i = 1; i < length; ++i){
+    for(size_t i = 1; i < length; ++i){
         key = a[i];
         j = i - 1;
         while(j >= 0 && key < a[j]){
